requirements.txt
openai-whisper
torch
transformers
pytesseract
